+++ demos/chinook/src/test/java/is/codion/framework/demos/chinook/model/TrackTableModelTest.java
@@ -43,7 +43,7 @@ public final class TrackTableModelTest {
 						.selectSingle(Album.TITLE.equalTo("Master Of Puppets"));
 
 		TrackTableModel trackTableModel = new TrackTableModel(connectionProvider);
-		trackTableModel.conditionModel()
+		trackTableModel.queryModel().conditionModel()
 						.setEqualOperand(Track.ALBUM_FK, masterOfPuppets);
 
 		trackTableModel.refresh();