+++ demos/world/src/test/java/is/codion/framework/demos/world/model/CountryTableModelTest.java
@@ -37,7 +37,7 @@ public final class CountryTableModelTest {
 	void fillCountryReport() throws ReportException {
 		try (EntityConnectionProvider connectionProvider = createConnectionProvider()) {
 			CountryTableModel tableModel = new CountryTableModel(connectionProvider);
-			tableModel.conditionModel().conditionModel(Country.CODE).operands().equal().set("ISL");
+			tableModel.queryModel().conditionModel().conditionModel(Country.CODE).operands().equal().set("ISL");
 			tableModel.refresh();
 			tableModel.selectionModel().selectedIndex().set(0);
 			tableModel.fillCountryReport(new ProgressReporter<String>() {