+++ framework/model/src/main/java/is/codion/framework/model/EntityModel.java
@@ -90,7 +90,7 @@ public interface EntityModel<M extends EntityModel<M, E, T>, E extends EntityEdi
 	/**
 	 * Adds the given detail model to this model, a side effect if the detail model contains
 	 * a table model is that it is configured so that a query condition is required for it to show
-	 * any data, via {@link EntityTableModel#conditionRequired()}.
+	 * any data, via {@link EntityQueryModel#conditionRequired()}.
 	 * Note that each detail model is associated with the first foreign key found referencing this models entity.
 	 * @param detailModels the detail models to add
 	 * @throws IllegalArgumentException in case no foreign key exists between the entities involved
@@ -100,7 +100,7 @@ public interface EntityModel<M extends EntityModel<M, E, T>, E extends EntityEdi
 	/**
 	 * Adds the given detail model to this model, a side effect if the detail model contains
 	 * a table model is that it is configured so that a query condition is required for it to show
-	 * any data, via {@link EntityTableModel#conditionRequired()}.
+	 * any data, via {@link EntityQueryModel#conditionRequired()}.
 	 * Note that the detail model is associated with the first foreign key found referencing this models entity.
 	 * @param detailModel the detail model
 	 * @return the resulting {@link ForeignKeyDetailModelLink}
@@ -111,7 +111,7 @@ public interface EntityModel<M extends EntityModel<M, E, T>, E extends EntityEdi
 	/**
 	 * Adds the given detail model to this model, a side effect if the detail model contains
 	 * a table model is that it is configured so that a query condition is required for it to show
-	 * any data, via {@link EntityTableModel#conditionRequired()}
+	 * any data, via {@link EntityQueryModel#conditionRequired()}
 	 * Specify the foreign key in case the detail model is based on an entity which contains multiple foreign keys to the
 	 * same master entity.
 	 * @param detailModel the detail model
@@ -123,7 +123,7 @@ public interface EntityModel<M extends EntityModel<M, E, T>, E extends EntityEdi
 	/**
 	 * Adds the given detail model to this model, a side effect if the detail model contains
 	 * a table model is that it is configured so that a query condition is required for it to show
-	 * any data, via {@link EntityTableModel#conditionRequired()}
+	 * any data, via {@link EntityQueryModel#conditionRequired()}
 	 * @param detailModelLink the {@link DetailModelLink} to add
 	 * @param <L> the {@link DetailModelLink} type
 	 * @return the {@link DetailModelLink}