+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityModelTest.java
@@ -42,7 +42,7 @@ public final class SwingEntityModelTest
 		SwingEntityModel employeeModel = new SwingEntityModel(Employee.TYPE, departmentModel.connectionProvider());
 		employeeModel.editModel().refreshComboBoxModels();
 		departmentModel.addDetailModel(employeeModel, Employee.DEPARTMENT_FK).active().set(true);
-		employeeModel.tableModel().conditionRequired().set(false);
+		employeeModel.tableModel().queryModel().conditionRequired().set(false);
 
 		return departmentModel;
 	}